requirements.txt
amqp==5.0.9
arrow==1.2.1
asgiref==3.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
blessed==1.19.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.0
Deprecated==1.2.13
Django==4.0
django-picklefield==3.0.1
django-q==1.3.9
ffmpeg-python==0.2.0
future==0.18.2
hiredis==2.0.0
ipython==7.30.1
jedi==0.18.1
kombu==5.2.3
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
psutil==5.9.0
psycopg2==2.9.2
ptyprocess==0.7.0
Pygments==2.10.0
pymediainfo==5.1.0
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
traitlets==5.1.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
